[metrics][user] Move metrics to their own package
