+++ user/metric/metric.go
@@ -1,4 +1,4 @@
-package main
+package metric
 
 import (
 	"github.com/prometheus/client_golang/prometheus"
@@ -6,17 +6,17 @@ import (
 )
 
 var (
-	requestCreate = "create"
-	requestRead   = "read"
-	requestUpdate = "update"
-	requestDelete = "delete"
+	RequestCreate = "create"
+	RequestRead   = "read"
+	RequestUpdate = "update"
+	RequestDelete = "delete"
 
-	requestSuccess = promauto.NewCounterVec(prometheus.CounterOpts{
+	RequestSuccess = promauto.NewCounterVec(prometheus.CounterOpts{
 		Name: "user_request_success_total",
 		Help: "The total number of successful requests",
 	}, []string{"request_type"})
 
-	requestFailure = promauto.NewCounterVec(prometheus.CounterOpts{
+	RequestFailure = promauto.NewCounterVec(prometheus.CounterOpts{
 		Name: "user_request_failure_total",
 		Help: "The total number of failed requests",
 	}, []string{"request_type", "error_code"})
